fix(CI) Align GitHub workflows, remove old scripts (#622)
* fix(CI) Align GitHub workflows, remove old scripts


* fix(build) Fix npm coverage for the new build
