--- .github/workflows/node.js.yml
@@ -1,59 +1,0 @@
-# This workflow will do a clean install of node dependencies, build the source code and run tests across different versions of node
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
-
-name: Build
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-
-jobs:
-  build:
-
-    strategy:
-      matrix:
-        node-version: [12.x, 14.x]
-        os: [ubuntu-latest, windows-latest, macOS-latest]
-        # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
-
-    runs-on: ${{ matrix.os }}
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v1
-      with:
-        node-version: ${{ matrix.node-version }}
-    
-    - name: use cache 
-      uses: actions/cache@v2
-      with:
-          path: |
-            node_modules
-            */*/node_modules
-          key: ${{ runner.os }}-${{ matrix.node_version }}-${{ hashFiles('**/package-lock.json') }}
-
-    - run: npm ci
-    - run: npm run build --if-present
-    - run: npm test
-
-  publish:
-    needs: build
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - name: Use Node.js 14.x
-      uses: actions/setup-node@v1
-      with:
-        node-version: 14.x
-    
-    - run: npm ci
-    - run: npm run build --if-present
-    - run: |
-          npm run pkgstamp
-          npm config set //registry.npmjs.org/:_authToken ${NPM_TOKEN}
-          npx lerna bootstrap
-          npx lerna exec -- npm publish --ignore-scripts --tag=unstable 2>&1
-      env:
-          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
