--- .github/workflows/npm-publish.yml
@@ -1,47 +1,0 @@
-# This workflow will run tests using node and then publish a package to GitHub Packages when a release is created
-# For more information see: https://help.github.com/actions/language-and-framework-guides/publishing-nodejs-packages
-
-name: Release
-
-on:
-  release:
-    types: [published]
-
-jobs:
-  publish:
-    name: Publish to npm
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
-        with:
-          node-version: 14
-      - name: build and publish
-        run: |
-          npm config set //registry.npmjs.org/:_authToken ${NPM_TOKEN}
-          npx lerna bootstrap
-          npm run pkgset ${{ github.event.release.tag_name }}
-          npx lerna exec -- npm publish --ignore-scripts --tag=latest 2>&1
-        env:
-          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
-
-      - name: create PR to increment version
-        id: cpr
-        uses: peter-evans/create-pull-request@v3
-        with:
-          base: master
-          commit-message: 'chore(actions): publish ${{ github.event.release.tag_name }} to npm'
-          committer: GitHub <noreply@github.com>
-          author: ${{ github.actor }} <${{ github.actor }}@users.noreply.github.com>
-          signoff: true
-          branch: cicero-publish-${{ github.event.release.tag_name }}
-          delete-branch: true
-          title: 'chore(actions): publish ${{ github.event.release.tag_name }} to npm'
-          body: |
-            # Increment Versions
-
-            Update the package.json version numbers after publishing to NPM.
-          assignees: ${{ github.actor }}
-          reviewers: ${{ github.actor }}
-          draft: false
