--- .travis/fix_github_https_repo.sh
@@ -1,47 +1,0 @@
-#/bin/bash
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-# 
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-#-- Script to automate https://help.github.com/articles/why-is-git-always-asking-for-my-password
-
-REPO_URL=`git remote -v | grep -m1 '^origin' | sed -Ene's#.*(https://[^[:space:]]*).*#\1#p'`
-if [ -z "$REPO_URL" ]; then
-  echo "-- ERROR:  Could not identify Repo url."
-  echo "   It is possible this repo is already using SSH instead of HTTPS."
-  exit
-fi
-
-USER=`echo $REPO_URL | sed -Ene's#https://github.com/([^/]*)/(.*).git#\1#p'`
-if [ -z "$USER" ]; then
-  echo "-- ERROR:  Could not identify User."
-  exit
-fi
-
-REPO=`echo $REPO_URL | sed -Ene's#https://github.com/([^/]*)/(.*).git#\2#p'`
-if [ -z "$REPO" ]; then
-  echo "-- ERROR:  Could not identify Repo."
-  exit
-fi
-
-NEW_URL="git@github.com:$USER/$REPO.git"
-echo "Changing repo url from "
-echo "  '$REPO_URL'"
-echo "      to "
-echo "  '$NEW_URL'"
-echo ""
-
-CHANGE_CMD="git remote set-url origin $NEW_URL"
-`$CHANGE_CMD`
-
-echo "Success"
